Changed the Dictionary in ConfDocument to use a composite key, rather than have to concatenate section and key on each lookup. This should help to reduce string allocations.
This change also made it so that a ConfDocument can no longer be searched with a single already concatenated string. Updated the one remaining use of that, the geta and getg commands, to parse out section and key prior to getting a setting.

Note: Even with this change, there are still string allocations in places where the config is read, such as the AuthFile module. It would be even better to allow searching by ReadOnlySpan<char> instead of string, but unfortunately that would require a custom Dictionary.