Added a null check to FreqManager.
I encountered a null reference exception while developing the matchmaking modules. However, I have not replicated the issue. Further investigation will be needed. It is likely a race condition between a player discconnecting (Network module's SendThread) and the mainloop thread. The FreqManager module still needs to be revisted with regards to threading.