Added the Peer module and updated functionality to use it, including: - The ?arena command. Also, updated to support chat output. - The ?z and ?az commands. - The ?find command. - ArenaManager redirect logic. - Directory publishing logic. - Network ping handler logic
The peer protocol sends arena names and player names very frequently. Allocating strings would likely create a lot of garbage that would need to be collected. Therefore, instead of allocating strings, made the logic use spans and pooled objects with character arrays. However, there is a need to search a collection by string. Use of a string keyed Dictionary or HashSet would require a string allocation. Therefore, added an implementation of the trie data structure. This will be useful in the future to spannify other parts of the server such as ConfigManager and CommandManager.