A complete overhaul of sized sends in the Network module.
Prior to this, the sized send logic closely matched that of ASSS. However, the ASSS logic has problems and limitations:
- If the game client requests to cancel a sized send, there's a race condition with the possiblity that the ISizedSendData object gets used after it's removed. In ASSS this is a use after free of the memory that was for a sized_send_data object. Also, the client could be sent sized data packets after the sized cancellation ACK (0x00 0x0C) is sent.
- The logic to queue more data is done using a mainloop timer. This means the call to request data, which may perform file I/O, is done on the mainloop thread. Also, since it's a timer, the throughput (max transfer rate) is limited by how frequently the timer is set to run.

Changed to use a thread dedicated to managing sized sends. The thread uses a producer-consumer queue. Items are added to the queue when:
- a sized send is added
- a sized send is requested to be cancelled (either by the game client requesting it or because the Player is being disconnected)
- an ACK is received for a sized send chunk

The thread handles:
- requesting data from the sender
- queuing sized chunk packets
- notifying the sender when a transfer is complete
- watching for cancellations
- sending sized cancellation ACK packets (0x00 0x0C)

For Players being disconnected, the SendThread will cancel any remaining sized sends and wait for the SizedSendThread to complete them before freeing the Player.

The new logic is also better when it comes to locks.
- The old logic held onto the global player data lock (read) for the entirety of queuing sized data since it needed to loop through the players collection. The new logic does not hold onto the global player data lock for nearly as long, as it does not loop over the players collection.
- The old logic used the olmtx when accessing the sized send list. The new logic uses a separate lock object just for the sized send fields of ConnData. Now, olmtx is only used when writing to the outgoing queues.

Added the ability to control how reliable callbacks are invoked: on the mainloop thread (the default), on a thread pool thread, or synchronously. Did not expose it on INetwork, but could in the future.

Also, updated the lagout logic to not allocate a string for the disconnect chat message.