Fix a race condition in how the Network module calls client connection handelrs.
Change the Network module to execute client connection handler methods on the mainloop thread. This helps guarantee they are executed in the proper order. That is, if it's handling received packets and also disconnecting the connection at the same time, the HandlePacket method calls will always get executed before the Disconnected method gets called.