Shutdown logic to be even more graceful.
Change how the mainloop shuts down. Now, when the mainloop is told to stop, the work queue is no longer marked as not accepting more additions. Instead, that is done on Unload. This means the mainloop will break out of its loop, but can decide to process more work items later in the shutdown process.

Update the Network module to fully clean up all connections. The buffers in their queues are also cleaned up. This includes buffers in outgoing reliable queues which may have reliable callbacks. The reliable callbacks are called, including ones that run on the mainloop (see the preceding note about mainloop changes). Additionally, Player objects are freed. Also, fixed an issue with sized send objects not getting retuned to their pool.

Update the BillingUdp module to disconnect gracefully on shutdown. It now waits for the entire disconnect sequence to complete. This includes: sending the B2S disconnect packet (0x03) reliably with a callback, the subsequent client connection disconnect request, and waiting for confirmation that the client connection has been disconnected.