Eliminate unnecessary string allocations when reading config settings in the following modules: Balls, ClientSettings, and LegalShip.
Fix an issue in the Balls module where it would try to clear the spawns array (TryReset) before it was allocated. This scenario was highly unlikely, the server getting shutdown before ArenaAction.Create was called on a new arena. Changed how it works, so that the array is always allocated, and in such a way that the per-arena data object can be reused without ever having to reallocate the array.