Fix a deadlock issue in the Network module.
Remove aquiring a connection's OutLock to call the SendRaw method, it shouldn't be necessary.

This fixes a deadlock issue due to locks not being entered in the proper order. The CorePacket_SyncRequest method would enter a connection's OutLock before the global PlayerData lock is entered by the SendRaw method. This was opposite to all other flows, which enter the global PlayerData lock before entering a connection's OutLock.

Thank you to plushmonkey for raising the issue and providing the trace.