Fix an issue where the Persist module would cause spikes.
The logic to sync all data was holding onto the global player lock and global arena lock for too long, causing the rest of the server to have to wait. To fix this, the locks are only held when populating temporary collections that identify the players and arenas to process. The locks are not held while the actual sync logic is processing. This should be ok, since the state of the players and arenas should not be able to complete a transition without having the persist thread, the same thread that's doing the full sync, process a request.

Thank you to plushmonkey for the player bots that have been useful to stress test the server, and for the help in identifying the cause.