Use the 0x39 (Small) and 0x3A (Large) Continuum position packets.
Note: This does not include batching them together yet. That will be done in a future change, probably to the Network module.

Fix an issue where the safe zone logic was never occuring because the old position data was already updated. Modify it to include not just entering, but also exiting a safe zone.

Modify weapon range logic to also consider the player's screen resolution when a weaspon is being fired.