Fix bugs with the nexthold command.
Give mods the proper permissions to private message the command to others.

Fix an issue where removing a hold with the nexthold command would not unset the target player's playing state.

Fix issues where the nexthold command was outputting the player name of the player executing the command rather than the target player.