Update TeamVersusMatch position packet filtering.
The logic already filtered position packets such that only players in the same match can see each other. Update it to allow players not in a match to only see other players that are also not in a match. This allows for having public play in the same arena that a team versus match is being held in. Spectators can still see all players. A future update will be needed to filter position packets for spectators to only receive position packets for the match that they are viewing (perhaps by who they are spectating).