Redesign of IPeriodicRewardPoints.
The old design was overly complicated. The GetRewardPoints method's IReadOnlyDictionary would have its enumerator boxed, causing an unnecessary allocation.

The new design simplifies it so that the IPeriodicRewardPoints implementation does all the work of gathering data for each team, forgoing the need for the need to pass team data.

Also, the logic to increment a player's flag points purposely never awards points to players in spectator mode or in a safe zone. This is so that it should keep points data in sync with what clients expect. The Periodic:IncludeSpectators and Periodic:IncludeSafeZones settings are still included. However, they have different meaning than in ASSS. They allow control over whether player counts should include spectators and players in safe zones, when calculating points to reward each team.