Overhaul of arena state machine logic.
Previously, a mainloop timer was used to periodically processes arena states. This timer would loop through all arenas, checking whether it could progress the arena further through its lifecycle of states. Much of the work performed in these steps is done asynchronously.

This change replaces the mainloop timer with a new mechanism, the StateMachineDirector. Upon a state change, the director can have the next stage of logic executed on the desired thread (e.g. mainloop thread).

This has the benefit of:
- One less timer that takes the global arena data lock for writing. So, likely less lock contention.
- Much faster processing of arena states. When one step completes, the next is queued up to be processed ASAP. This means faster arena creation, recycling, etc.
- Ability to use async/await instead of having to poll for the completion of Tasks.