Add a way for clients to specify which features they support.
Use the 2 unused bytes of the 444 field in the login packet as a bitset that tells which features a client supports. This is useful for bots that use VIE authentication to tell the server which features it supports. For example, the ZeroBot can tell the server it supports batch positions and watch damage. Update the server logic to read the data and use it to toggle the associated features.