Add the abilty to filter bricks.
- Add the IBricksAdvsior interface.
- Update the bricks module to use IBricksAdvsior.
- Update the MatchFocus module to filter bricks.

This allows for matches to not interfere with one another when they are in the same physical box. It's not a perfect solution, since there really isn't one. Spectators will see the bricks for the match that they are spectating. There is no way to tell clients to clear bricks, so they will continue to see those bricks if they swtich to another match. Players in a match will only see the bricks for the match they are in. If a brick was placed before a player becomes focused on the match (e.g., subs in) or enters the arena after the brick was placed, the player will not see those bricks. If a player was spectating a match and saw a brick placed, that player will continue to see the brick if they are placed in a match that uses the box that they were previously spectating.