Refactor to use MemoryExtensions.Split instead of StringUtils.GetToken
Update the CommandManager module to not trim whitespace from the end of parameters. This way, trailing whitespace will be sent when sending to the billing server. For example, this can affect the results of the ?find command ("?find X-Demo" vs "?find X-Demo " with a space at the end). Requested by X-Demo.