Use ConcurrentQueue<T> instead of ConcurrentBag<T> for the object pooling mechanism in which the pooled objects know which pool they originated from.
This is an attempt to workaround a System.OverflowException with using ConcurrentBag which happens at System.Collections.Concurrent.ConcurrentBag`1.WorkStealingQueue.LocalPush(T item, Int64& emptyToNonEmptyListTransitionCount)
The closest reported issue that could be found was:
https://github.com/dotnet/runtime/issues/114817
but it has no resolution.

I orginally used ConcurrentBag over ConcurrentQueue since it's mostly the same thread that is renting and returning objects, though not always. I assume the ConcurrentQueue will not have this issue since it does not have the thread local logic of ConcurrentBag.